get skies from SpyroWorldViewer working; `special_skies` -> `levels`
convert UVW from Spyro level OBJs to RGB "Vertex Colors" / "Color Attribute" in a Blender object. Blender can't read UVW data fully, it just drops the W value and keeps the UV, so we have to read the file outside of Blender and then apply the vertex colors that way.